separate drug settings in separate file
